Create SVcf2bed-DELnINS.pl
A short perl script that reads a VCF from STDIN containing SVs (tag SVTYPE) and outputs a bed file. Only SVTYPEs of DEL or INS are retained. For deletions the start and end coords in bed file correspond to size of DEL and for insertions the the start and end correspond to the INS breakpoint. Bed file coords begin with zero not 1